Extracted classes from ID3TagCreator :recycle:
